mgr/cephadm: move _check_host to serve.py
